image2Pdf/requirements.txt
Pillow==8.4.0  # PIL library
reportlab==3.6.1